fix(luaengine): resolve all C_LUA compilation and linking errors
- Upgrade sol2 3.2.3 → 3.3.0 with patches for noexcept guards,
  function_pointer_type, and optional<T&>::emplace
- Switch to C++17 globally (autotools can't do per-target -std=)
- Add -D_GNU_SOURCE to CXXFLAGS (required with _XOPEN_SOURCE=700)
- Delete vs/unistd.h (empty shim that shadowed system <unistd.h>)
- Add #include <unistd.h> to 9 asio .ipp files
- Create debug_bridge.h/.cpp: centralized extern declarations and
  no-op stubs for CBreakpoint API, replacing 23 #if 0 blocks across
  12 files
- Fix instrumentation_router.cpp: add #include config.h before #if C_LUA
  guard (entire file was compiling to empty object with 0 symbols)
- Add pc98_cdl.cpp to misc/Makefile.am (was never compiled)
- Add linker group (-Wl,--start-group/--end-group) to resolve circular
  dependency: hardware → luaengine → libmisc
- Fix includes: dosbox.h + mem.h before debug_inc.h for Bitu/PhysPt,
  logging.h for DEBUG_ShowMsg, <climits> for PATH_MAX
- Fix SegValue(cs) → SegValue(SegNames::cs) type mismatch
- Fix std::min type deduction with int64_t casts
- Fix ::GetAddress → GetAddress (extern forward decl)
- Build verified: 144MB ELF64 binary, 0 errors, exit code 0

+++ src/luaengine/cheat_engine.cpp
@@ -279,13 +279,13 @@ void Cheat::applyIncreaseDecrease() {
     // Clamp to valid range based on size
     switch (size_) {
         case LuaEngineRamSearch::WatchSize::BYTE_1:
-            new_value = std::max(0LL, std::min(255LL, new_value));
+            new_value = std::max(int64_t(0), std::min(int64_t(255), new_value));
             break;
         case LuaEngineRamSearch::WatchSize::BYTE_2:
-            new_value = std::max(0LL, std::min(65535LL, new_value));
+            new_value = std::max(int64_t(0), std::min(int64_t(65535), new_value));
             break;
         case LuaEngineRamSearch::WatchSize::BYTE_4:
-            new_value = std::max(0LL, std::min(4294967295LL, new_value));
+            new_value = std::max(int64_t(0), std::min(int64_t(4294967295), new_value));
             break;
         case LuaEngineRamSearch::WatchSize::BYTE_8:
             // No clamping for 64-bit
